A push/pop capable vector

An R6 class that implements a persistent vector with push and pop operations.

Public fields

values

A vector of values

Methods

Public methods

pvector$new()

pvector$print()

pvector$length()

pvector$pop()

pvector$push()

pvector$clone()

Method new()

Create a pvector

Usage

pvector$new(values = c())

Arguments

values

A vector of values

Returns

A new pvector

Method print()

Print pvector

Usage

pvector$print()

Returns

NULL

Method length()

Get length of pvector

Usage

pvector$length()

Returns

An integer

Method pop()

Pop vector

Usage

pvector$pop()

Returns

Popped value

Method push()

Push values

Usage

pvector$push(pushed.values)

Arguments

pushed.values

A vector of values

Returns

NULL

Method clone()

The objects of this class are cloneable with this method.

Usage

pvector$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.